requirements.txt
boto3==1.36.16
botocore==1.36.16
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
jinja2==3.1.5
jmespath==1.0.1
markupsafe==3.0.2
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyyaml==6.0.2
requests==2.32.3
s3transfer==0.11.2
six==1.17.0
urllib3==2.3.0
